% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Navigate directories and print contents}
\usage{
cd(x, ...)
}
\arguments{
\item{x}{Your directory or empty string}

\item{...}{Other arguments passed to methods}
}
\description{
Navigate directories and print contents
}
