% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Apply a function to some data, and then write the result}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{vector of something to chunk}

\item{n}{number of chunks}
}
\value{
chunks of data
}
\description{
Via chunking
}
\examples{
chunk(1:10000, 5) # 5 chunks
}
