% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countp.R
\name{countp}
\alias{countp}
\title{Counts rows and calculates proportions}
\usage{
countp(df, col, ...)
}
\arguments{
\item{df}{a dataframe}

\item{col}{column to count on}

\item{...}{other arguments passed on to methods}
}
\description{
Counts rows and calculates proportions
}
