% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_churn}
\alias{cpp_churn}
\title{The number of periods since x was positive (churn).}
\usage{
cpp_churn(x)
}
\arguments{
\item{x}{input numeric vector}
}
\value{
periods since last positive value
}
\description{
The number of periods since x was positive (churn).
}
