% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprint.R
\name{dprint}
\alias{dprint}
\title{Prints the directories of your current wd}
\usage{
dprint(max_files = 11, max_nchar = 20, ...)
}
\arguments{
\item{max_files}{The maximum number of files to show per bin}

\item{max_nchar}{The maximum length of a file name to show}

\item{...}{Other arguments passed to methods}
}
\description{
Prints the directories of your current wd
}
