% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive_ctas.R
\name{generate_hive_ctas}
\alias{generate_hive_ctas}
\title{Generate Hive CTAS statement from R dataframe}
\usage{
generate_hive_ctas(dataframe, comment = NA, table = "random",
  partitition = NULL)
}
\arguments{
\item{dataframe}{data}

\item{comment}{the description of each column}

\item{table}{schema.table name}

\item{partitition}{quoted partitions to include}
}
\value{
a query to execute in hive
}
\description{
Generate Hive CTAS statement from R dataframe
}
\examples{
generate_hive_ctas(data.frame(
  chr = letters[1:5],
  lgl = c(TRUE, FALSE, TRUE, FALSE, TRUE),
  dbl = runif(5),
  int = 15L:19L,
  time = Sys.time(),
  date = Sys.Date()
))
}
