% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_quarter.R
\name{hc_quarter}
\alias{hc_quarter}
\title{Returns JavaScript formatting date to quarterly values}
\usage{
hc_quarter(ds, just_quarter = TRUE, additional_js = NULL)
}
\arguments{
\item{ds}{The datetime on the x-axis}

\item{just_quarter}{Just return the quarter..for X axis label}

\item{additional_js}{Character string to pass on as JavaScript}
}
\description{
Returns JavaScript formatting date to quarterly values
}
