% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive_ctas.R
\name{hive_ctas}
\alias{hive_ctas}
\title{Generate Hive CTAS statement from R dataframe}
\usage{
hive_ctas(dataframe, comment = NULL, table = "random",
  partitition = NULL, tbl_properties = NULL)
}
\arguments{
\item{dataframe}{data}

\item{comment}{the description of each column}

\item{table}{schema.table name}

\item{partitition}{quoted partitions to include}

\item{tbl_properties}{named list of properties to give}
}
\value{
a create table statement executable in hive
}
\description{
Generate Hive CTAS statement from R dataframe
}
\examples{
hive_ctas(data.frame(
  chr = letters[1:5],
  lgl = c(TRUE, FALSE, TRUE, FALSE, TRUE),
  dbl = runif(5),
  int = 15L:19L,
  time = Sys.time(),
  date = Sys.Date()
),
partitition = "date"
)
}
