% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Move some files or rename them}
\usage{
move(x, dir_to)
}
\arguments{
\item{x}{files or dirs that you want to move}

\item{dir_to}{directory/filename to move files to}
}
\description{
Just a wrapper for file.rename
}
