% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remove_rows.R
\name{remove_db_rows}
\alias{remove_db_rows}
\title{Conditionally remove rows from a database table}
\usage{
remove_db_rows(db, table, column, vals)
}
\arguments{
\item{db}{a database connection}

\item{table}{the table to modify}

\item{column}{the column to apply filters on}

\item{vals}{the values to use in the filter}
}
\description{
Allows you to drop specified rows from a database rather than deleting the whole table
}
