% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_files_in_wd.R
\name{return_files_in_wd}
\alias{return_files_in_wd}
\title{Returns a list of files}
\usage{
return_files_in_wd(show_hidden = FALSE)
}
\arguments{
\item{show_hidden}{Whether to show hidden files}
}
\description{
The list is formatted by the object type it is interpreted as
}
