% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ipsum.R
\name{theme_ipsum}
\alias{theme_ipsum}
\title{An attempted port of `theme_ipsum` into `Highcharter`}
\usage{
theme_ipsum(...)
}
\arguments{
\item{...}{arguments passed to methods}
}
\description{
Supplies a new theme inspired by theme_ipsum
}
