% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_data.R
\name{write_chunks}
\alias{write_chunks}
\title{Apply a function to some data, and then write the result}
\usage{
write_chunks(dataframe, fun, dir, chunks)
}
\arguments{
\item{dataframe}{data}

\item{fun}{function to apply}

\item{dir}{directory to write data to}

\item{chunks}{a discrete number of chunks to iterate over}
}
\description{
Via chunking
}
